requirements.txt
streamlit
python-dotenv
crewai
crewai-tools
pandas
matplotlib
